[Feature #162686097] Add PROGATE_EXCEPTIONS setting to config file to permit detection of custom flask_jwt_extended errors
