+++ app/api/v2/__init__.py
@@ -14,4 +14,4 @@ from flask_restful import Api
 bp = Blueprint('v2', __name__)
 api_bp = Api(bp)
 
-from app.api.v2 import models, auth, incidents, decorators
+from app.api.v2 import models, auth, incidents, errors
